Setting globTtl to 2 seconds only in dev env
The globTtl setting (that commands when to do file scans for changed files) is now set to 2 seconds for the "dev" environment and set to null (i.e. cache forever) for other environments.
This should result in a performance boost for production machines (especially if they receive less than 1 request every 2 seconds).